refactor: simplify log viewer state
