--- src/tui/hooks/useFilters.js
@@ -1,20 +1,0 @@
-import { useState, useRef, useEffect } from 'react';
-import { DEFAULT_FILTERS } from '../ui-utils/entry-utils.js';
-
-export const useFilters = () => {
-  const [filters, setFilters] = useState(DEFAULT_FILTERS);
-  const [showFilterView, setShowFilterView] = useState(false);
-  const filtersRef = useRef(filters);
-
-  useEffect(() => {
-    filtersRef.current = filters;
-  }, [filters]);
-
-  return {
-    filters,
-    setFilters,
-    showFilterView,
-    setShowFilterView,
-    filtersRef,
-  };
-};
