--- src/tui/hooks/useTraceBuffer.js
@@ -1,8 +1,0 @@
-import { useState } from 'react';
-
-export const useTraceBuffer = () => {
-  const [buffer, setBuffer] = useState(null);
-  const [entries, setEntries] = useState([]);
-
-  return { buffer, setBuffer, entries, setEntries };
-};
